Remove `$debug` parameter from scripts and styles
We originally added an optional `$debug` parameter to the functions used to enqueue scripts and styles in order to make them efficiently testable without corrupting constants. Unfortunatately, many people copy-paste these methods as-is without understanding the point of the parameter and/or using tests properly.

To prevent poor copy-pasta code, we're removing the parameter and instead adding filters to both the script and style enqueues to continue to support testability.

Fixes #54