The tree method should run over each item
While refactoring the tree method, the structure of the tree
became unclear. The former implementation was running over each
set of items passing it to the processing function. Consequently
this was also what was being tested :(. This update fixes the
implementation so that it is now compatible with it actual needed
behavior. Next up will be updating the tests so that they are
ensuring proper behavior for the *needed* functionality, not the
one-off functionality that was happening.